+++ uefi-macros/src/lib.rs
@@ -211,3 +211,57 @@ pub fn entry(args: TokenStream, input: TokenStream) -> TokenStream {
     };
     result.into()
 }
+
+/// Builds a `CStr8` literal at compile time from a string literal.
+///
+/// This will throw a compile error if an invalid character is in the passed string.
+///
+/// # Example
+/// ```
+/// # use uefi_macros::cstr8;
+/// assert_eq!(cstr8!("test").to_bytes_with_nul(), [116, 101, 115, 116, 0]);
+/// ```
+#[proc_macro]
+pub fn cstr8(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    let input: LitStr = parse_macro_input!(input);
+    let input = input.value();
+    match input
+        .chars()
+        .map(u8::try_from)
+        .collect::<Result<Vec<u8>, _>>()
+    {
+        Ok(c) => {
+            quote!(unsafe { ::uefi::CStr8::from_bytes_with_nul_unchecked(&[ #(#c),* , 0 ]) }).into()
+        }
+        Err(_) => syn::Error::new_spanned(input, "invalid character in string")
+            .into_compile_error()
+            .into(),
+    }
+}
+
+/// Builds a `CStr16` literal at compile time from a string literal.
+///
+/// This will throw a compile error if an invalid character is in the passed string.
+///
+/// # Example
+/// ```
+/// # use uefi_macros::cstr16;
+/// assert_eq!(cstr16!("test €").to_u16_slice_with_nul(), [116, 101, 115, 116, 32, 8364, 0]);
+/// ```
+#[proc_macro]
+pub fn cstr16(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    let input: LitStr = parse_macro_input!(input);
+    let input = input.value();
+    match input
+        .chars()
+        .map(|c| u16::try_from(c as u32))
+        .collect::<Result<Vec<u16>, _>>()
+    {
+        Ok(c) => {
+            quote!(unsafe { ::uefi::CStr16::from_u16_with_nul_unchecked(&[ #(#c),* , 0 ]) }).into()
+        }
+        Err(_) => syn::Error::new_spanned(input, "invalid character in string")
+            .into_compile_error()
+            .into(),
+    }
+}